# Conflicts:
#	db/java-waiter.mv.db
#	src/main/java/com/waiter/javawaiter/dish/controller/DishController.java
#	src/main/java/com/waiter/javawaiter/dish/mapper/DishMapper.java
#	src/main/java/com/waiter/javawaiter/dish/model/Dish.java
#	src/main/java/com/waiter/javawaiter/dish/repository/DishRepository.java
#	src/main/java/com/waiter/javawaiter/dish/service/DishService.java
#	src/main/java/com/waiter/javawaiter/dish/service/DishServiceImpl.java
#	src/main/java/com/waiter/javawaiter/employee/service/EmployeeServiceImpl.java
#	src/main/java/com/waiter/javawaiter/order/controller/OrderController.java
#	src/main/java/com/waiter/javawaiter/order/dto/OrderDto.java
#	src/main/java/com/waiter/javawaiter/order/dto/OrderForAdminDto.java
#	src/main/java/com/waiter/javawaiter/order/dto/OrderShortDto.java
#	src/main/java/com/waiter/javawaiter/order/mapper/OrderMapper.java
#	src/main/java/com/waiter/javawaiter/order/model/Order.java
#	src/main/java/com/waiter/javawaiter/order/repository/OrderRepository.java
#	src/main/java/com/waiter/javawaiter/order/service/OrderService.java
#	src/main/java/com/waiter/javawaiter/order/service/OrderServiceImpl.java
#	src/main/resources/schema.sql
#	src/test/java/com/waiter/javawaiter/dish/DishControllerTest.java
#	src/test/java/com/waiter/javawaiter/dish/DishServiceImplTest.java
#	src/test/java/com/waiter/javawaiter/employee/EmployeeServiceImplTest.java
#	src/test/java/com/waiter/javawaiter/order/OrderControllerTest.java
#	src/test/java/com/waiter/javawaiter/order/OrderServiceImplTest.java

+++ src/main/java/com/waiter/javawaiter/order/service/OrderServiceImpl.java
@@ -1,5 +1,10 @@
 package com.waiter.javawaiter.order.service;
 
+import com.waiter.javawaiter.comment.mapper.CommentMapper;
+import com.waiter.javawaiter.comment.model.Comment;
+import com.waiter.javawaiter.comment.repository.CommentRepository;
+import com.waiter.javawaiter.dish.dto.DishForOrderDto;
+import com.waiter.javawaiter.dish.mapper.DishMapper;
 import com.waiter.javawaiter.dish.model.Dish;
 import com.waiter.javawaiter.dish.repository.DishRepository;
 import com.waiter.javawaiter.employee.repository.EmployeeRepository;
@@ -14,12 +19,12 @@ import com.waiter.javawaiter.order.model.Order;
 import com.waiter.javawaiter.order.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -27,9 +32,12 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class OrderServiceImpl implements OrderService {
     private final OrderRepository orderRepository;
-    private final DishRepository dishRepository;
+    private final DishMapper dishMapper;
+    private final CommentRepository commentRepository;
     private final EmployeeRepository employeeRepository;
     private final OrderMapper orderMapper;
+    private final CommentMapper commentMapper;
+    private final DishRepository dishRepository;
 
     @Override
     public OrderDto create(Long employeeId, OrderShortDto order, LocalDateTime localDateTime) {
@@ -37,14 +45,16 @@ public class OrderServiceImpl implements OrderService {
         var employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
         var thisOrder = orderMapper.toOrder(order);
-        List<Optional<Dish>> dishes = order.getDishes().stream().map(dishRepository::findById).toList();
-        thisOrder.setDishes(dishes.stream().flatMap(Optional::stream).filter(Objects::nonNull).toList());
+        decreaseQuantity(thisOrder.getDishes());
         thisOrder.setEmployee(employee);
-        thisOrder.setStatus(Status.CREATED);
+        thisOrder.setCreationTime(localDateTime);
         thisOrder.setBillTime(null);
         thisOrder.setTotal(null);
         log.info("Добавлен заказ: {}", thisOrder);
-        return orderMapper.toOrderDto(orderRepository.save(thisOrder));
+        saveCommentsForDishes(order.getDishes());
+        var orderForReturn = orderMapper.toOrderDto(orderRepository.save(thisOrder));
+        orderForReturn.setDishes(order.getDishes());
+        return orderForReturn;
     }
 
     @Override
@@ -52,45 +62,47 @@ public class OrderServiceImpl implements OrderService {
         log.info("updateOrder({}, {}, {})", employeeId, order, orderId);
         isAcceptable(employeeId, orderId);
         var thisOrder = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Заказ не найден"));
+        if (thisOrder.getBillTime() != null || thisOrder.getTotal() != null) {
+            throw new ValidationViolationException("Заказ закрыт, нельзя изменить данные");
+        }
         if (order.getGuests() != null) {
             thisOrder.setGuests(order.getGuests());
         }
         if (!order.getDishes().isEmpty()) {
-            List<Dish> dishes = order.getDishes().stream().map(dishRepository::findById)
-                    .flatMap(Optional::stream)
-                    .filter(Objects::nonNull).toList();
+            List<Dish> dishes = order.getDishes().stream().map(dishMapper::toDish).toList();
             thisOrder.setDishes(dishes.stream()
                             .peek(dish -> {
                                 if (thisOrder.getDishes().stream().anyMatch(dishes::contains)) {
                                     thisOrder.getDishes().add(dish);
                 }
             }).collect(Collectors.toList()));
+            decreaseQuantity(thisOrder.getDishes());
         }
         if (order.getCreationTime() != thisOrder.getCreationTime()) {
             throw new ValidationViolationException("Нет доступа к изменению времени заказа");
         }
-        log.info("Обновлен заказ: {}", thisOrder);
-        return orderMapper.toOrderDto(orderRepository.save(thisOrder));
+        var orderForReturn = orderMapper.toOrderDto(orderRepository.save(thisOrder));
+        saveCommentsForDishes(orderForReturn.getDishes());
+        List<DishForOrderDto> dishes = orderForReturn.getDishes();
+        orderForReturn.setDishes(getCommentsForDishes(dishes));
+        log.info("Обновлен заказ: {}", orderForReturn);
+        return orderForReturn;
     }
 
     @Override
     public void deleteById(Long employeeId, Long orderId) {
         log.info("deleteOrderById({}, {})", employeeId, orderId);
+        var order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Заказ не найден"));
+        if (order.getBillTime() != null || order.getTotal() != null) {
+            throw new ValidationViolationException("Заказ закрыт, нельзя изменить данные");
+        }
+        List<Long> dishes = order.getDishes().stream().map(Dish::getDishId).toList();
+        for (Long dish: dishes) {
+            commentRepository.deleteCommentsByDishId(dish);
+        }
         isAcceptable(employeeId, orderId);
         orderRepository.deleteById(orderId);
         log.info("Заказ с идентификатором {} удален пользователем с идентификатором {}", orderId, employeeId);
-    }
-
-    @Override
-    public void deleteOrders(Long employeeId) {
-        log.info("deleteOrders({})", employeeId);
-        var employee = employeeRepository.findById(employeeId)
-                .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
-        if (!employee.getIsAdmin()) {
-            throw new AccessViolationException("Нет доступа");
-        }
-        orderRepository.deleteAll();
-        log.info("Список заказов очищен пользователем с идентификатором: {}", employeeId);
     }
 
     @Override
@@ -98,30 +110,86 @@ public class OrderServiceImpl implements OrderService {
         log.info("getOrderById({}, {})", employeeId, orderId);
         isAcceptable(employeeId, orderId);
         Order thisOrder = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Заказ не найден"));
-        log.info("Возвращено блюдо: {}, по запросу пользователя с идентификатором {}", orderId, employeeId);
-        return orderMapper.toOrderDto(thisOrder);
+        List<DishForOrderDto> dishes = thisOrder.getDishes().stream()
+                .map(dishMapper::toDishForOrderDto).collect(Collectors.toList());
+        var order = orderMapper.toOrderDto(thisOrder);
+        order.setDishes(getCommentsForDishes(dishes));
+        log.info("Возвращено блюдо: {}, по запросу пользователя с идентификатором {}", order, employeeId);
+        return order;
     }
 
     @Override
-    public List<OrderDto> getOrders(Long employeeId) {
+    public List<OrderDto> getOrders(Long employeeId, int offset, int limit) {
         log.info("Запрос на получение списка заказов от пользователя: {}", employeeId);
+        PageRequest page = PageRequest.of(offset, limit);
         var employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
-        if (!employee.getIsAdmin()) {
-            throw new AccessViolationException("Нет доступа к изменению заказа");
+        List<OrderDto> orders;
+        if (employee.getIsAdmin()) {
+            orders = orderRepository.findAll(page).stream().map(orderMapper::toOrderDto).collect(Collectors.toList());
+        } else {
+            orders = orderRepository.findAllByEmployee(employee, page).stream()
+                    .map(orderMapper::toOrderDto).collect(Collectors.toList());
         }
-        List<Order> orders = orderRepository.findAll();
-        log.info("Возвращён список заказов: {}", orders);
-        return orders.stream().filter(order -> !Objects.equals(order.getEmployee().getEmployeeId(), employeeId))
-                .map(orderMapper::toOrderDto).collect(Collectors.toList());
+        for (OrderDto o: orders) {
+            o.setDishes(getCommentsForDishes(o.getDishes()));
+        }
+        log.info("Возвращён список заказов: {}, для пользователя: {}", orders, employee);
+        return orders;
     }
 
     private void isAcceptable(Long employeeId, Long orderId) {
+        log.info("isAcceptable: {}, {}", employeeId, orderId);
         var employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
         var thisOrder = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Заказ не найден"));
         if (!thisOrder.getEmployee().getEmployeeId().equals(employeeId) || !employee.getIsAdmin()) {
             throw new AccessViolationException("Нет доступа к изменению заказа");
+        }
+    }
+
+    private List<Dish> updateStatus(Order order) {
+        return order.getDishes().stream().peek(dish -> {
+            if (order.getCreationTime().equals(order.getCreationTime().plusMinutes(1L))) {
+                dish.setStatus(Status.IN_PROGRESS);
+            }
+            if (LocalDateTime.now().equals(order.getCreationTime().plusMinutes(dish.getTimeLimit() + 1L))) {
+                dish.setStatus(Status.DONE);
+            }
+            if (order.getCreationTime().equals(order.getCreationTime().plusMinutes(dish.getTimeLimit() + 1L + 10L))) {
+                dish.setStatus(Status.SERVED);
+            }
+        }).collect(Collectors.toList());
+    }
+
+    private List<DishForOrderDto> getCommentsForDishes(List<DishForOrderDto> dishes) {
+        List<DishForOrderDto> list = new ArrayList<>();
+        for (DishForOrderDto dish: dishes) {
+            Comment comment = commentRepository.findByDishId(dish.getDishId());
+            dish.setComment(commentMapper.toCommentShortDto(comment));
+            list.add(dish);
+        }
+        return list;
+    }
+
+    private void saveCommentsForDishes(List<DishForOrderDto> dishes) {
+        for (DishForOrderDto dish: dishes) {
+            if (dish.getComment() == null) {
+                continue;
+            }
+            commentRepository.save(new Comment(dish.getComment().getCommentId(),
+                    dish.getDishId(), dish.getComment().getComment()));
+        }
+    }
+
+    private void decreaseQuantity(List<Dish> dishes) {
+        for (Dish dish: dishes) {
+            if (dish.getQuantity() != 0) {
+                dish.setQuantity(dish.getQuantity() - 1);
+                dishRepository.save(dish);
+            } else {
+                throw new ValidationViolationException("Блюда " + dish + " нет в наличии");
+            }
         }
     }
 }
